Refactor following PR review
