+++ applications/launchpad/gui-react/src/App.tsx
@@ -8,10 +8,8 @@ import HomePage from './pages/home'
 import { loadDefaultServiceSettings } from './store/settings/thunks'
 import './styles/App.css'
 
-import TBotManager from './TBotManager'
-import { selectTBotQueue } from './store/tbot/selectors'
-
 import useMiningSimulator from './useMiningSimulator'
+import TBotContainer from './containers/TBotContainer'
 
 const AppContainer = styled.div`
   background: ${({ theme }) => theme.background};
@@ -23,7 +21,6 @@ const AppContainer = styled.div`
 const App = () => {
   const themeConfig = useAppSelector(selectThemeConfig)
   const dispatch = useAppDispatch()
-  const tbotQueue = useAppSelector(selectTBotQueue)
 
   dispatch(loadDefaultServiceSettings())
 
@@ -35,7 +32,7 @@ const App = () => {
     <ThemeProvider theme={themeConfig}>
       <AppContainer>
         <HomePage />
-        <TBotManager messages={tbotQueue} />
+        <TBotContainer />
       </AppContainer>
     </ThemeProvider>
   )
