+++ applications/launchpad/gui-react/src/components/TBot/HelpComponents/styles.ts
@@ -2,5 +2,5 @@ import styled from 'styled-components'
 
 export const StyledTextContainer = styled.div`
   display: flex;
-  margin-bottom: 24px;
+  margin-bottom: ${({ theme }) => theme.spacingVertical(2)};
 `
