+++ applications/launchpad/gui-react/src/components/TBot/TBotPrompt/styles.ts
@@ -76,23 +76,6 @@ export const StyledCloseIcon = styled.div`
   z-index: 3;
 `
 
-export const StyledMessage = styled(animated.div)`
-  max-width: 385px;
-  height: fit-content;
-  margin-left: ${({ theme }) => theme.spacingHorizontal(0.6)};
-  margin-right: ${({ theme }) => theme.spacingHorizontal(0.6)};
-  margin-bottom: ${({ theme }) => theme.spacingHorizontal(0.6)};
-  background-color: ${({ theme }) => theme.background};
-  opacity: 1;
-  border-radius: ${({ theme }) => theme.borderRadius(2)};
-  box-shadow: ${({ theme }) => theme.shadow24};
-  padding: 40px;
-  color: ${({ theme }) => theme.primary};
-  &:last-child {
-    margin-bottom: 0;
-  }
-`
-
 export const DotsContainer = styled.div`
   display: flex;
   flex-direction: row;
